Renamed StateDB to TrackedDB and move to its own module
